Added Sweeping Strike for #38

+++ Core/src/main/java/util/Settings.java
@@ -11,8 +11,8 @@ public class Settings {
 	public static final int FPS = 30;
     
 	// seconds to frames
-	public static int seconds(int sec){
-		return sec * FPS;
+	public static int seconds(double sec){
+		return (int) (sec * FPS);
 	}
 	public static int framesToSeconds(int frames){
 		return frames / FPS;
